pyproject.toml
[project]
name = "oai-agents-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "logfire>=3.24.0",
    "openai-agents[litellm]>=0.1.0",
    "python-dotenv>=1.1.1",
]
